Update uploadFilesModel.js
arreglar error en el que se trataban de enviar dos facturas al mismo tiempo y se caia el baquend

+++ models/uploadFilesModel.js
@@ -4,25 +4,28 @@ import { fileURLToPath } from 'url'
 import path from 'path'
 import { getDownloadURL, uploadBytes, ref } from 'firebase/storage' // Importar las funciones necesarias de Firebase Storage
 
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
+const filename = fileURLToPath(import.meta.url)
+const dirname = path.dirname(filename)
 
 export class UploadFilesModel {
   static async uploadFiles (files, dir) {
     try {
       const { archivo } = files
-      const uploadPath = path.join(__dirname, '../temp', archivo.name)
+      const timestamp = Date.now()
+      const uploadPath = path.join(dirname, '../temp', `${timestamp}_${archivo.name}`)
       await archivo.mv(uploadPath)
-      const storageRef = ref(storage, `${dir}/${Date.now()}${archivo.name}`)
+      const storageRef = ref(storage, `${dir}/${timestamp}_${archivo.name}`)
       await uploadBytes(storageRef, fs.readFileSync(uploadPath))
       const downloadURL = await getDownloadURL(storageRef)
       fs.unlinkSync(uploadPath)
       return downloadURL
     } catch (error) {
       const { archivo } = files
-      const uploadPath = path.join(__dirname, '../temp', archivo.name)
-      fs.unlinkSync(uploadPath)
-      return (error)
+      const uploadPath = path.join(_dirname, '../temp', `${Date.now()}${archivo.name}`)
+      if (fs.existsSync(uploadPath)) {
+        fs.unlinkSync(uploadPath)
+      }
+      return error
     }
   }
 }
